users-service/requirements.txt
flask==3.0.3
gunicorn==21.2.0
psycopg2-binary==2.9.9
kafka-python==2.0.2